pyproject.toml
[project]
name = "qwen3-agent"
version = "0.1.0"
description = "Email search agent training with Qwen3 14B using OpenPipe ART"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core ART library for RL training
    "openpipe-art>=0.5.1",
    "tblib>=3.0.0",
    
    # Data processing
    "datasets>=3.4.1",
    "diskcache>=5.6.3",
    "huggingface-hub>=0.29.3",
    "kaggle>=1.5.12",
    "langchain-core>=0.3.51",
    "mail-parser>=3.15.0",
    "pandas>=1.3.0",
    "polars>=1.27.1",
    
    # ML/AI (Core - CPU compatible)
    "openai>=1.0.0",
    "litellm>=1.65.0.post1",
    "transformers>=4.50.3",
    "torch>=2.5.1",
    "peft",
    "torchtune",
    "torchao",
    
    # RL Training (SB3)
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.0.0",
    
    # GPU-specific (Linux only) - using platform markers
    # These will be automatically skipped on macOS
    "vllm; sys_platform == 'linux'",
    "unsloth; sys_platform == 'linux'",
    "unsloth-zoo; sys_platform == 'linux'",
    
    # Utils
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "tenacity>=8.0.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.62.0",
    "pydantic>=2.0.0",
    "pytest>=8.3.5",
    "setuptools",
    
    # Visualization & Monitoring
    "matplotlib>=3.10.1",
    "wandb>=0.18.0",
    
    # Notebook support
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    
    # Cloud & AWS
    "skypilot[runpod]>=0.8.1",
    "awscli>=1.38.1",
    
    # Performance
    "hf-transfer>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["qwen3_agent*"]

[dependency-groups]
dev = [
    "qwen3-agent",
]

